requirements.txt
google-cloud-bigquery==3.11.4
pandas-gbq==0.19.2

